refactor: address self-review nits from #1583 fix
- Share trackedSpawnDeadlineMs/trackedSpawnStartupDeadlineMs docs once via
  a new TrackedSpawnDeadlineOptions interface, extended by CompleteOptions
  and OpenSessionOpts instead of duplicating per-field comments.
- Move trackedSpawnDeadlines() out of config/selectors.ts into its own
  tracked-spawn-deadlines.ts — it isn't a ConfigSelector (its callers hold
  a config slice or a possibly-undefined NaxConfig, not the full NaxConfig
  ConfigSelector.select() expects), so it doesn't belong in the "Named
  ConfigSelector Registry" file.
- Dedupe the spawn-result test stubs (makeSpawnResult / makeWedgedSpawnResult)
  and the process.kill safety stub between spawn-client.test.ts and
  spawn-client-tracked-spawn-deadlines.test.ts into a shared
  _spawn-client-test-helpers.ts, following the existing
  test/integration/tdd/_tdd-test-helpers.ts underscore-prefixed local-helper
  convention.
